fix: connection creation to external URL
